+++ packages/terra-table/src/subcomponents/Cell.jsx
@@ -103,7 +103,7 @@ const propTypes = {
   onCellSelect: PropTypes.func,
 
   /**
-   * String that specifies the height of the cell. Any valid CSS value is accepted.
+   * String that specifies the min-height of the cell. Any valid CSS value is accepted.
    */
   height: PropTypes.string,
 };
@@ -281,7 +281,7 @@ function Cell(props) {
   const CellTag = isRowHeader ? 'th' : 'td';
 
   // eslint-disable-next-line react/forbid-dom-props
-  let cellContentComponent = <div className={cx('cell-content', theme.className)} style={{ height }}>{cellContent}</div>;
+  let cellContentComponent = <div className={cx('cell-content', theme.className)} style={{ 'min-height': height }}>{cellContent}</div>;
   // Render FocusTrap container when within a grid context
   if (isGridContext) {
     cellContentComponent = (