+++ packages/terra-time-input/src/_AccessibleInput.jsx
@@ -25,11 +25,16 @@ const propTypes = {
   isInvalid: PropTypes.bool,
   /** When true, setting isInvalid to true will also visually display the input as invalid. */
   showIsInvalid: PropTypes.bool,
+  /**
+   * String that labels the error message for screen readers.
+   */
+  ariaDescribedBy: PropTypes.string,
 };
 
 const defaultProps = {
   isInvalid: false,
   showIsInvalid: false,
+  ariaDescribedBy: null,
 };
 
 /**
@@ -54,6 +59,7 @@ function AccessibleInput(props) {
     disabled,
     isInvalid,
     showIsInvalid,
+    ariaDescribedBy,
     ...inputProps
   } = props;
 
@@ -81,8 +87,7 @@ function AccessibleInput(props) {
         aria-invalid={isInvalid}
         aria-labelledby={value ? labelId : undefined}
         aria-label={value ? undefined : label}
-        aria-describedby={descriptionId}
-
+        aria-describedby={ariaDescribedBy ? `${descriptionId} ${ariaDescribedBy}` : `${descriptionId}`}
       />
       <VisuallyHiddenText aria-live={TimeUtil.isMac() ? 'polite' : 'off'} id={descriptionId} text={description} />
     </>