+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* Changed
+  * Added `date-picker`, `time-input` and `date-time picker` tests to display new metadata.
+
 ## Unreleased
 
 ## 1.59.0 - (January 22, 2024)