+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Updated `date-input` examples to allow screen reader response for error message.
+
 * Changed
   * Updated `terra-table` test example to support large text data.