+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Added
+  * Updated `date-input` examples to allow screen reader response for error message.
+
+* Changed
+  * Updated `terra-table` test example to support large text data.
+
+## 1.59.0 - (January 22, 2024)
+
 * Fixed
   * Fixed font size in `terra-theme-provider` examples.
   * Fixed label value when we have 12 hours time format in `terra-time-input` examples.